Split SearchInput into a controller hook and presentational pieces
useSearchInputController owns all state and event logic (debounce-settle
machine, IME composition, paste resolution via lib/search-input-logic,
type switching, drawer state); SearchTypeHint and SearchDrawers hold the
hint chip and the lazy radical/handwriting drawers. SearchInput.tsx is now
a ~100-line presentational component wiring the three together. Behavior
is unchanged, guarded by the paste-inference and drawer E2E specs.
